// src/generated/client/worldmonitor/market/v1/service_client.ts
export interface GetSectorSummaryResponse {
  sectors: SectorPerformance[];
}

export interface SectorPerformance {
  symbol: string;
  name: string;
  change: number;
}